gcp-helper/requirements.txt
flask==3.0.0
google-cloud-compute==1.14.1
google-auth==2.23.4
gunicorn==21.2.0
